ShaderBuffers now use device local memory.
They are now uploaded in upload step.

+++ source/Core/Castor3D/Buffer/UniformBufferPool.cpp
@@ -24,8 +24,8 @@ namespace castor3d
 				, dst.makeTransferDestination() );
 			commandBuffer.copyBuffer( src
 				, dst
-				, uint32_t( size )
-				, uint32_t( offset ) );
+				, size
+				, offset );
 			dstSrcStage = dst.getCompatibleStageFlags();
 			commandBuffer.memoryBarrier( dstSrcStage
 				, flags
